# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CSE 341 Books and Authors API",
    "version": "1.0.0",
    "description": "Books and Authors Web Service for CSE 341 Week 02"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "summary": "Get all books",
        "tags": [
          "Books"
        ],
        "responses": {
          "200": {
            "description": "Books returned successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new book",
        "tags": [
          "Books"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book created successfully"
          },
          "400": {
            "description": "Invalid data, duplicate ID, or author not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "summary": "Get a book by ID",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom book ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book returned successfully"
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update a book",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom book ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book updated successfully"
          },
          "400": {
            "description": "Invalid data or author not found"
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a book",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom book ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Book deleted successfully"
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/authors": {
      "get": {
        "summary": "Get all authors",
        "tags": [
          "Authors"
        ],
        "responses": {
          "200": {
            "description": "Authors returned successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new author",
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Author created successfully"
          },
          "400": {
            "description": "Invalid data or duplicate author ID"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "summary": "Get an author by ID",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom author ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Author returned successfully"
          },
          "404": {
            "description": "Author not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update an author",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom author ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Author updated successfully"
          },
          "400": {
            "description": "Invalid author data"
          },
          "404": {
            "description": "Author not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete an author",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom author ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Author deleted successfully"
          },
          "400": {
            "description": "Author cannot be deleted because books reference this author"
          },
          "404": {
            "description": "Author not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BookInput": {
        "type": "object",
        "required": [
          "id",
          "authorId",
          "title",
          "publicationDate"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "b4"
          },
          "authorId": {
            "type": "string",
            "example": "a1"
          },
          "title": {
            "type": "string",
            "example": "A New Test Book"
          },
          "publicationDate": {
            "type": "string",
            "format": "date",
            "example": "2026-09-26"
          }
        }
      },
      "BookUpdate": {
        "type": "object",
        "required": [
          "authorId",
          "title",
          "publicationDate"
        ],
        "properties": {
          "authorId": {
            "type": "string",
            "example": "a1"
          },
          "title": {
            "type": "string",
            "example": "Updated Test Book"
          },
          "publicationDate": {
            "type": "string",
            "format": "date",
            "example": "2026-09-26"
          }
        }
      },
      "AuthorInput": {
        "type": "object",
        "required": [
          "id",
          "name",
          "birthYear"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "a4"
          },
          "name": {
            "type": "string",
            "example": "Fourth Author"
          },
          "birthYear": {
            "type": "integer",
            "example": 1985
          }
        }
      },
      "AuthorUpdate": {
        "type": "object",
        "required": [
          "name",
          "birthYear"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Updated Author"
          },
          "birthYear": {
            "type": "integer",
            "example": 1986
          }
        }
      }
    }
  },
  "tags": []
}